Fix #13609: read MKV sequentially with a 1 MB buffer on network shares
Revert the parallel cluster walk from #13629. On the reporter's share
(Synology DS1621xs+ over SMB, 901 MB file, 11 subtitle tracks) it made
parsing far worse, not better: beta13 ~5 s, beta14 132 s.

Frame-by-frame analysis of the reporter's two screen recordings splits
beta14 cleanly in half. For the first 65 s the progress bar never leaves
indeterminate - not one progress callback fired - so that is phase 1,
DiscoverClusters, which is strictly serial: seek to the next top-level
element, read two vints, seek again. The remaining 66 s is the 16-way
parallel walk. A serial pass costing the same wall clock as a 16-way
parallel one means per-read latency dominates completely and the
parallelism barely pays for the extra pass.

The rig that motivated #13629 pointed the wrong way for two reasons,
both biased against sequential reads:

  - it mounted CIFS with cache=none,actimeo=0 "to be wire-honest", which
    disables client-side read-ahead - the very mechanism that makes big
    sequential reads fast on a real SMB client. The rig's ceiling for
    64 KB sequential was 37 MB/s; the reporter's share does ~225 MB/s.
  - its server-side store was a container filesystem on page cache, so a
    random 4 KB read cost only a ~0.05 ms round-trip. On a NAS the same
    read also costs an array seek, and defeats read-ahead at both ends.

So the rig ranked "few bytes, many round-trips" above "many bytes, few
round-trips". Real hardware ranks them the other way by a wide margin:
on that share the whole 901 MB streams in about 4 s, so any strategy
trading bytes for seeks is losing before it starts. That is also why
5.1.0's sparse 4 KB walk was slow there in the first place - #13629
doubled down on the pattern that was the original bug.

The network buffer also goes 64 KB -> 1 MB. At 64 KB we already pull
~80% of the file over the wire, so a larger buffer costs almost no extra
bytes while cutting the round-trip count proportionally.

Kept from #13629: the ClusterReader extraction, and FindTrackStartInCluster
no longer collecting subtitle blocks as a side effect. The equivalence
test between the two walks goes away with the parallel walk; a smaller
test takes its place, guarding that the cluster walk still extracts
subtitles in playback order from the sample MKVs - I/O tuning must not
change what is parsed.
