+++ src/libse/Properties/AssemblyInfo.cs.template
@@ -26,4 +26,3 @@ using System.Runtime.InteropServices;
 
 // Make the internal types available in our testing assembly.
 [assembly: InternalsVisibleTo("Test")]
-[assembly: InternalsVisibleTo("LibSETests")]
